Factor out Schema DDL
