+++ project/Dependencies.scala
@@ -19,7 +19,7 @@ object Dependencies {
     val igluClient         = "0.5.0"
     val igluCore           = "0.3.0-M1"
     val processingManifest = "0.1.0-M4"
-    val schemaDdl          = "0.7.0"
+    val schemaDdl          = "0.8.0-M1"
     // Scala
     val cats               = "1.1.0"
     val catsEffect         = "1.0.0-RC2"
